Extract the running of Juvet to another module so it can be shared
